// web/remote-control/src/generated/types/WorkspaceSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DeviceSummary } from "./DeviceSummary";
import type { HostSnapshot } from "./HostSnapshot";
import type { ProjectSnapshot } from "./ProjectSnapshot";
import type { RecentAgentSessionSnapshot } from "./RecentAgentSessionSnapshot";
import type { TargetRef } from "./TargetRef";
import type { UsageSnapshot } from "./UsageSnapshot";

export type WorkspaceSnapshot = { revision: number, sequence: number, host: HostSnapshot, projects: Array<ProjectSnapshot>, active_target: TargetRef | null, usage: Array<UsageSnapshot>, recent_agent_sessions: Array<RecentAgentSessionSnapshot>, paired_devices: Array<DeviceSummary>, };